Refac: made config more OO
